# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorBody'
      success:
        type: boolean
    type: object
...
  request.CreateTagRequest:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
      slug:
        description: 可选，为空则自动生成
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  request.UpdateTagRequest:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
      slug:
        description: 可选，为空则保持不变
        maxLength: 50
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API 支持
    url: http://www.swagger.io/support
  description: 博客后端 API 服务，支持文章管理、配置文件认证等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 博客系统 API
  version: "1.0"
paths:
...
  /tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取标签列表
      tags:
      - tag
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建标签
      tags:
      - tag
  /tags/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: 无内容
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除标签
      tags:
      - tag
    get:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据ID获取标签
      tags:
      - tag
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新标签
      tags:
      - tag
  /tags/slug/{slug}:
    get:
      parameters:
      - description: 标签Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据Slug获取标签
      tags:
      - tag
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证令牌，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"